Day 16
Write and test a function which takes two arguments (a year and a month) and returns the number of days for the given month/year pair (while only February is sensitive to the year value, your function should be universal).
Convince the function to return None if its arguments don't make sense. Use a list filled with the months' lengths.